requirements.txt
python-telegram-bot==13.7
pyshorteners==1.0.1
validators==0.20.0
